% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{FindAllMarkers2}
\alias{FindAllMarkers2}
\title{Find markers for all groups}
\usage{
FindAllMarkers2(
  object,
  meta_data = NULL,
  group_column = NULL,
  replicate_column = NULL,
  method = "edger",
  verbosity = 1
)
}
\arguments{
\item{object}{Input object (of class Seurat, SingleCellExperiment, dgCMatrix, or matrix)}

\item{meta_data}{Data frame of meta data; can be NULL if object is Seurat; default is NULL}

\item{group_column}{String or integer indicating what column of meta_data to use as group indicator;
if NULL, then the function tries to find an appropriate column; default is NULL}

\item{replicate_column}{String or integer indicating what column of meta_data
to use as replicate indicator; default is NULL}

\item{method}{DE method to use; default is edger}

\item{verbosity}{Integer controlling how many messages the function prints;
0 is silent, 1 prints some messages, 2 prints more messages}
}
\value{
A data frame of results
}
\description{
This calls [findDE()] with compare = 'each_vs_rest' and filters and orders the results.
}
