% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{findDE}
\alias{findDE}
\title{Differential expression tests for single-cell RNA-seq data}
\usage{
findDE(
  object,
  meta_data = NULL,
  group_column = NULL,
  replicate_column = NULL,
  compare = "each_vs_rest",
  method = "edger",
  order_results = TRUE,
  verbosity = 1
)
}
\arguments{
\item{object}{Input object (of class Seurat, SingleCellExperiment, dgCMatrix, or matrix)}

\item{meta_data}{Data frame of meta data; can be NULL if object is Seurat or
SingleCellExperiment and cell identities are set; default is NULL}

\item{group_column}{String or integer indicating what column of meta_data to use as group indicator;
if NULL, the object cell identities are used; default is NULL}

\item{replicate_column}{String or integer indicating what column of meta_data
to use as replicate indicator; default is NULL}

\item{compare}{Specifies which groups to compare - see details; default is 'each_vs_rest'}

\item{method}{DE method to use, one of edger, deseq, limma; default is edger}

\item{order_results}{Whether to order the results by comparison, then by p-value, then
by test statistic. If FALSE, results will be ordered by comparison only, genes remain
in the order as in the input. Default is TRUE}

\item{verbosity}{Integer controlling how many messages the function prints;
0 is silent, 1 prints some messages, 2 prints more messages}
}
\value{
A data frame of results with the following columns
\itemize{
\item \strong{feature} the gene (as given by rownames of input)
\item \strong{ave_expr} average expression of gene (renamed method specific values; edger: logCPM, deseq: baseMean, limma: Amean)
\item \strong{log_fc} log fold-change (renamed method specific values; edger: logFC, deseq: log2FoldChange, limma: LogFC)
\item \strong{stat} test statistic (renamed method specific values; edger: F, deseq: stat, limma: lods)
\item \strong{pvalue} test p-value
\item \strong{padj} adjusted p-value (FDR)
\item \strong{rate1} detection rate in group one (fraction of cells with non-zero counts)
\item \strong{rate2} detection rate in group two (fraction of cells with non-zero counts)
\item \strong{group1} comparison group one
\item \strong{group2} comparison group two
}

The log fold-change values are estimates of the form log2(group1/group2) - details will depend on the method chosen.
}
\description{
Differential expression tests for single-cell RNA-seq data
}
\section{Details}{

Compare groups of cells using DESeq2, edgeR, or limma-trend.

There are multiple ways the group comparisons can be specified based on the compare
parameter. The default, \code{'each_vs_rest'}, does multiple comparisons, one per
group vs all remaining cells. \code{'all_vs_all'}, also does multiple comparisons,
covering all group pairs. If compare is set to a length two character vector, e.g.
\code{c('T-cells', 'B-cells')}, one comparison between those two groups is done.
To put multiple groups on either side of a single comparison, use a list of length two.
E.g. \code{compare = list(c('cluster1', 'cluster5'), c('cluster3'))}.
}

\examples{
\donttest{
# Example data
counts <- DElegate::pbmc$counts
meta_data <- DElegate::pbmc$meta_data

# Use matrix and meta data as input and perform all "this group vs rest" comparisons
de_res <- findDE(object = counts,
                 meta_data = meta_data,
                 group_column = 'celltype')

# Same with Seurat object as input
s <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta_data)
Seurat::Idents(s) <- s$celltype
de_res <- findDE(object = s)
}

}
